Add process tags to DSM (#7775)
## Summary of changes

- add a hash of process tags to the base hash of pathway checkpoint
hashes
- write the process tags in DSM payloads(so that the backend can resolve
the right service name)

Since this makes us query the process tags during the tracer
initialization, I removed the Lazy mechanism that I had put in place
previously, because there would be nothing lazy about it anymore.
This means that values that are available only later in the app life
(like IIS entrypoint for instance) are not accessible.

still controlled by the same feature flag as in #7651 

This is the dotnet equivalent of
https://github.com/DataDog/dd-trace-java/pull/8836

## Reason for change

[AIDM-195]

## Implementation details

## Test coverage

## Other details
<!-- Fixes #{issue} -->


<!--  ⚠️ Note:

Where possible, please obtain 2 approvals prior to merging. Unless
CODEOWNERS specifies otherwise, for external teams it is typically best
to have one review from a team member, and one review from apm-dotnet.
Trivial changes do not require 2 reviews.

MergeQueue is NOT enabled in this repository. If you have write access
to the repo, the PR has 1-2 approvals (see above), and all of the
required checks have passed, you can use the Squash and Merge button to
merge the PR. If you don't have write access, or you need help, reach
out in the #apm-dotnet channel in Slack.
-->


[AIDM-195]:
https://datadoghq.atlassian.net/browse/AIDM-195?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ