Add process tags to dynamic instrumentation snapshots (#7839)
## Summary of changes

add a process tags field next to where the service name is added in the
snapshot json.

dotnet version of https://github.com/DataDog/dd-trace-java/pull/8779

## Reason for change

AIDM-195

## Implementation details

I wanted to inject the setting from
https://github.com/DataDog/dd-trace-dotnet/blob/79b266b0c7a329eae91584356a1cfc7c5599c1ee/tracer/src/Datadog.Trace/Debugger/DebuggerFactory.cs#L30
but looking at the call chain to bring it to the snapshot creator, it
was quite tedious and messy.
Another option was to put it in the debugger settings, but it felt out
of place as those settings are specific to the live debugger, not a
repeat of global tracer settings.
In the end I decided to inject _a little bit_ so that I could specify
the value in tests.

## Test coverage

just made sure the test that checks json correctness includes this

## Other details
<!-- Fixes #{issue} -->


<!--  ⚠️ Note:

Where possible, please obtain 2 approvals prior to merging. Unless
CODEOWNERS specifies otherwise, for external teams it is typically best
to have one review from a team member, and one review from apm-dotnet.
Trivial changes do not require 2 reviews.

MergeQueue is NOT enabled in this repository. If you have write access
to the repo, the PR has 1-2 approvals (see above), and all of the
required checks have passed, you can use the Squash and Merge button to
merge the PR. If you don't have write access, or you need help, reach
out in the #apm-dotnet channel in Slack.
-->