Avoid allocating and preparing CI Visibility stuff in startup hot path (#7821)
## Summary of changes

Avoid allocating and preparing CI Visibility stuff in startup hot path

## Reason for change

Currently CI Visibility uses a bunch of lazy settings, but then we
immediately instantiate those right at the start of the app to decide
_if_ we should initialize CI Visibility 😅

This just extracts out some of that work out to minimize the work if CI
Vis is not enabled.

This shaves ~50ms off startup

## Implementation details

Make things lazy that can be, make things non-lazy that can't be!

Concretely:
- `Enabled` shouldn't be lazy because we check it immediately.
Therefore:
- Extract it from `TestOptimizationSettings` (as it doesn't need to be
there anyway)
- Make `InternalEnabled()` testable by moving to a helper (not done
here, but could be)
  - Calculate it immediately in the ctor
- Make `CIEnvironmentVariables` calculation lazy, as we don't need it if
we're not in CI Vis

## Test coverage

Same coverage right now.