Add process tags to client-side stats (#7791)
## Summary of changes

adding process tags at the ClientStatsPayload level.
See
[proto](https://github.com/DataDog/datadog-agent/blob/main/pkg/proto/datadog/trace/stats.proto#L52-L55)
in the agent

## Reason for change

AIDM-195

## Implementation details

## Test coverage

piggybacked on existing tests

## Other details

I realized that dotnet doesn't write the service name at the top level,
but only in the buckets. Doesn't seem to be a big issue, but there could
be a win here by writing the service in the buckets only when there is
an override.



<!--  ⚠️ Note:

Where possible, please obtain 2 approvals prior to merging. Unless
CODEOWNERS specifies otherwise, for external teams it is typically best
to have one review from a team member, and one review from apm-dotnet.
Trivial changes do not require 2 reviews.

MergeQueue is NOT enabled in this repository. If you have write access
to the repo, the PR has 1-2 approvals (see above), and all of the
required checks have passed, you can use the Squash and Merge button to
merge the PR. If you don't have write access, or you need help, reach
out in the #apm-dotnet channel in Slack.
-->

---------

Co-authored-by: Andrew Lock <andrew.lock@datadoghq.com>