Add request-id as header to lambda start/end invocation (#7835)
## Summary of changes
Adds` lambda-runtime-aws-request-id` header in start invocation and end
invocation requests from tracer to extension. Assigns header to the aws
request id of the span.

## Reason for change
Allows the lambda extension to identify which start/end invocation
requests belong to which invocation if multiple Lambda requests are
being served concurrently.

## Implementation details
Uses the LambdaContext from the `IInvocationRequest proxyinstance` to
get the aws-request-id in `SendStartInvocation` and adds it to request
headers in `WriteRequestHeaders`. Saves the aws-request-id as part of
the state passed from `OnDelegateBegin `to `EndInvocationAsync`.
`EndInvocationAsync` passes the state down to `GetEndInvocationRequest`,
which adds it to the request headers.

Updates the LambdaCommonTests and LambdaRequestBuilderTests to match new
function signatures with state passed down and to test for the
assignment to the lambda-runtime-aws-request-id.

## Test coverage
Added unit test in LambdaRequestBuilderTests to check for `
lambda-runtime-aws-request-id` header and value in end invocation
request. Manually tested for headers in start and end invocation using
custom dd-trace-dotnet layer
`arn:aws:lambda:sa-east-1:425362996713:layer:dotnet-request-id-header-rithika:2`
and custom lambda extension
`arn:aws:lambda:sa-east-1:425362996713:layer:Debug-extension-rithika:1`
which prints the start and end invocation headers. Sample trace with
logs of the headers
[here](https://support-admin.us1.prod.dog/admin/switch_handle_get/org_id/1543931?next_url=%2Fapm%2Ftrace%2F219302768028734365%3FgraphType%3Dservice_map%26shouldShowLegend%3Dtrue%26spanID%3D4621724130062486128%26timeHint%3D1763567215307.0002%26trace%3D2193027680287343654621724130062486128%26traceQuery%3D)

## Other details
<!-- Fixes #{issue} -->


<!--  ⚠️ Note:

Where possible, please obtain 2 approvals prior to merging. Unless
CODEOWNERS specifies otherwise, for external teams it is typically best
to have one review from a team member, and one review from apm-dotnet.
Trivial changes do not require 2 reviews.

MergeQueue is NOT enabled in this repository. If you have write access
to the repo, the PR has 1-2 approvals (see above), and all of the
required checks have passed, you can use the Squash and Merge button to
merge the PR. If you don't have write access, or you need help, reach
out in the #apm-dotnet channel in Slack.
-->

---------

Co-authored-by: Lucas Pimentel <lucas.pimentel@datadoghq.com>