[Azure Functions] Fix Azure Functions NuGet script for Linux and macOS (#7864)
## Summary of changes

Fixed the `Build-AzureFunctionsNuget.ps1` script to work on Linux and
macOS in addition to Windows.

## Reason for change

The script only worked on only Windows.

## Implementation details

- Added OS detection logic to select the appropriate Nuke build script
(`build.ps1` for Windows, `build.sh` for Linux/macOS)
- Changed all hardcoded backslash paths to forward slashes which work in
all platforms
- Updated paths in verbose output messages for consistency

## Test coverage

Tested manually on Windows and Linux. This script is only used to build
local dev versions of the nuget package, not production code.

## Other details

n/a