[Profiler] Generate heap snapshot (#7681)
## Summary of changes
Allow the generation of heap snapshots that are sent to the backend

## Reason for change
For memory leak workflow, we would like to compare the surviving objects
count to identify the leaky ones. These heap snapshots will be
transformed into class histograms to look like what is generated by
Java.

## Implementation details
The .NET runtime allows the creation of "gcdump" thanks to dotnet-gcdump
tool. Behind the scene, an induced GC is triggered when an event pipe
session is created with the right keyword/verbosity. Undocumented events
are then emitted by the runtime at the end of the GC - read
https://chnasarre.medium.com/net-gcdump-internals-fcce5d327be7?source=friends_link&sk=3225ff119458adafc0e6935951fcc323
for more details.

## Test coverage
Test added as needed

## Other details
<!-- Fixes #{issue} -->


<!--  ⚠️ Note:

Where possible, please obtain 2 approvals prior to merging. Unless
CODEOWNERS specifies otherwise, for external teams it is typically best
to have one review from a team member, and one review from apm-dotnet.
Trivial changes do not require 2 reviews.

MergeQueue is NOT enabled in this repository. If you have write access
to the repo, the PR has 1-2 approvals (see above), and all of the
required checks have passed, you can use the Squash and Merge button to
merge the PR. If you don't have write access, or you need help, reach
out in the #apm-dotnet channel in Slack.
-->