Add fix for not re-reporting telemetry when there are "empty" dynamic config changes (#7796)
## Summary of changes

A fix for #7724 to handle telemetry reporting in dynamic config "reset"
scenarios

## Reason for change

The system tests for #7724 were failing in some dynamic configuration
scenarios. Specifically, the tests were sending remote config _without_
any configuration values "i.e. 'reset to use defaults'" and were waiting
a telemetry update. However, we never sent it, because there was "no
telemetry to record".

Note that we _did_ correctly apply the new configuration, we just didn't
report the telemetry correctly, primarily due to limitations in the
telemetry protocol. This PR adds a fix for that, and will be merged into
#7724.

## Implementation details

The solution is to "remember" the telemetry from the default mutable
configuration values, _without_ any dynamic sources, and "replay" this
telemetry when we update telemetry. This feels kind of hacky, but it's
something I suspected we might need to do, and had been avoiding up to
this point because we do a "full reconfigure" anyway.

## Test coverage

Added a specific unit test that mimics the behaviour of the system-test
(i.e. an "empty" dynamic config response) and confirms the telemetry is
recorded as expected

## Other details

https://datadoghq.atlassian.net/browse/LANGPLAT-819

Part of a config stack

- https://github.com/DataDog/dd-trace-dotnet/pull/7522
- https://github.com/DataDog/dd-trace-dotnet/pull/7525
- https://github.com/DataDog/dd-trace-dotnet/pull/7530
- https://github.com/DataDog/dd-trace-dotnet/pull/7532
- https://github.com/DataDog/dd-trace-dotnet/pull/7543
- https://github.com/DataDog/dd-trace-dotnet/pull/7544
- https://github.com/DataDog/dd-trace-dotnet/pull/7721
- https://github.com/DataDog/dd-trace-dotnet/pull/7722
- https://github.com/DataDog/dd-trace-dotnet/pull/7695
- https://github.com/DataDog/dd-trace-dotnet/pull/7723
- https://github.com/DataDog/dd-trace-dotnet/pull/7724
- https://github.com/DataDog/dd-trace-dotnet/pull/7796 👈

Unlike other PRs in the stack, I'll merge this directly into #7724 to
fix the tests there, just thought I'd keep this separate for easier
reviewing