Add benchmark for two scopes (#7869)
## Summary of changes

- Adds a new micro benchmark

## Reason for change

We have a benchmark for creating a span, and a scope, but I'm working on
some perf improvements, and I want to be able to see the improvements
specifically when there are two scopes in a trace.

## Implementation details

Add a new simple benchmark

## Test coverage

Not really, this is so that there's something to compare against in my
update!