[Profiler] Update liveheap handle limit (#7872)
## Summary of changes
Make the number of monitored live objects customizable via environment
variable and change the default value to 4096

## Reason for change
The current 1024 maximum was to small especially in case of memory leak

## Implementation details
Add the corresponding env var helpers in the configuration and use it in
the `LiveObjectsProvider`

## Test coverage
Tests have been added

## Other details
<!-- Fixes #{issue} -->


<!--  ⚠️ Note:

Where possible, please obtain 2 approvals prior to merging. Unless
CODEOWNERS specifies otherwise, for external teams it is typically best
to have one review from a team member, and one review from apm-dotnet.
Trivial changes do not require 2 reviews.

MergeQueue is NOT enabled in this repository. If you have write access
to the repo, the PR has 1-2 approvals (see above), and all of the
required checks have passed, you can use the Squash and Merge button to
merge the PR. If you don't have write access, or you need help, reach
out in the #apm-dotnet channel in Slack.
-->