Benchmark improvements (#7868)
## Summary of changes

- Remove the obsolete `[AgentFilter]`
- Allow running some benchmarks on master only (not PRs)

## Reason for change

The `[AgentFilter]` is left over from before we moved to benchmarking
platform, so delete it.

Additionally some benchmarks don't seem like they need to be run _every_
PR, so added separate categories to allow controlling this.

> [!IMPORTANT]
> This PR _doesn't_ change which benchmarks we run on PR vs master, it
just _allows_ us to do that easily by adding/removing a category

## Implementation details

- Delete the old filter
- Add new categories - add them to every benchmark currently tagged with
tracer/ASM
- Change how the filtering works in `RunBenchmarks` Nuke target to work
with this
- Requires benchmarking platform update:
  - https://github.com/DataDog/benchmarking-platform/pull/215

## Test coverage

This has the same coverage as today