Update .gitlab/one-pipeline.locked.yaml (#7876)
Automatic PR. Updates .gitlab/one-pipeline.locked.yaml

Note: The value matches this one:
https://gitlab.ddbuild.io/DataDog/apm-reliability/libdatadog-build/-/jobs/1259973862
Created from this job:
https://gitlab.ddbuild.io/DataDog/apm-reliability/libdatadog-build/-/pipelines/84225039

Changes:

https://github.com/DataDog/libdatadog-build/commit/50244c43322cb4e7231f6d275395b7764af067ac

[Config Registry] Create Supported In CI Job to Support V2 Local Files
(https://github.com/DataDog/libdatadog-build/pull/152)
Previously, we had a CI job that passed in data from the tracers to the
FPD backend to update the Supported In range on the FPD for new release
tags. We are now introducing a new v2 format for the data coming from
the tracers. Since the data is in a new v2 format, this CI job handles
the comparisons between local files for both v1 and v2 formats. Tracers
can migrate to using this CI job without implementing the v2 format of
the local files.