Add more cleanup to benchmarks (#7865)
## Summary of changes

- Update the JetBrains SelfApi package to latest version
- Don't run background tasks like remote config, telemetry, discovery
service in benchmarks
- Cleanup "global" tracers after benchmark

## Reason for change

Benchmarks should be isolated to keep them consistent, but unfortunately
a lot of our code has `Tracer.Instance` and `Security.Instance`
scattered around. That means we need to set up _global_ state for some
of the benchmarks. Currently, we're never cleaning up that global state,
which could introduce variance (flake) in the benchmarks).

> Note that this actually doesn't impact us in _most_ cases, only when
using the inprocess tool chain, which is primarily used for when we're
using dotmemory etc locally

Similarly, new `Tracer` instances start a bunch of background tasks,
such as remote configuration, telemetry, discovery service. These all
have the possibility of skewing results and introducing flake.
 
## Implementation details

- Cleanup any resources (e.g. `Tracer` instances, but anything
disposable) in `[GlobalCleanup]`
- Ensure we _remove_ and replace the global `Tracer` instances in
cleanup
- Disable services that start background jobs (discovery service,
telemetry, remote config)
- Create a helper for creating a tracer with default benchmarking
configuration

## Test coverage

This is the test - we may see some variation from the current results,
but hopefully the results will be more stable subsequently

## Other details

I _wish_ we could not use `Tracer.Instance` (or `Security.Instance`, or
`Iast.Instance`) at all. It would make all of this much easier to reason
about and reduce the chance of accidental cross-talk. That's something
we'll move towards slowly, for now this is probably the best we can do.