Stop testing .NET 8 on PRs (#7881)
## Summary of changes

Removes .NET 8 from the standard test matrix for PRs

## Reason for change

We recently added .NET 10 to the standard test matrix for PRs. As
running more tests means more chance of flake, we agreed that removing
.NET 8 to bring us back to the same level of testing makes sense.

- We chose to remove .NET 8 vs .NET 9, as there are CLR fixes in .NET 9
that aren't in .NET 8
- Both .NET 8 and .NET 9 have the same amount of official support
remaining
- Both .NET 8 and .NET 9 hit the same code paths in our code (which is
all .NET 6+)

Also, to be clear, we still run the full suite of TFMs (.NET Core 3.0+)
on `master`. We also run the full suite of TFMs on PRs if there are
"risky" modifications (modifications to integrations), or if the
`run_all_test_frameworks` variable is set.

## Implementation details

Removed .NET 8 from the "on PRs" frameworks

## Test coverage

Technically the _same_ coverage, but a bit less comprehensive on PRs