Only pull docker images if we need to (#7910)
## Summary of changes

Only pulls docker images if we need to

## Reason for change

Docker hub appears to be throwing `500` errors, and even though the
images are pre-pulled, we're calling `docker pull` which then fails.
Instead, we can only pull if we _need_ to, as we only pre-pull
unchanging images anyway.

## Implementation details

Use docker manifest to check if it exists, then docker pull if not

## Test coverage

This is the test

## Other details

Through this to AI, so if it doesn't work, don't blame me