Fix `DatadogLoggingScope` sometimes injecting incorrect values for `ILogger` logs injection (#7880)
## Summary of changes

Fixes a bug where `ILogger` logs injection injects the wrong values when
a customer changes service/env/version using manual instrumentation.

## Reason for change

This has been a bug for a _long_ time. Basically we cache a
`DatadogLoggingScope` instance on app startup, and then we _always_ use
it for injection. If the settings change, we didn't refresh this value,
and so would inject the wrong settings. A unit test confirmed this
behaviour.

## Implementation details

I opted for a simple fix for now, especially seeing as this hasn't been
flagged as a bug in escalations etc

- Store the current `MutableSettings` for the trace on the
`DatadogLoggingScope`
- When calling `AddScope`, check if the _current_ `MutableSettings`
match the ones for the provided scope
  - If they do, use the cached logging scope
  - If they don't, create a new one, update the cached scope

Overall, this works, as long as changes are one way, i.e. we don't have
much overlap of traces with different MutableSettings. This probably
won't be a problem, hence why I took this approach, but it feels a bit
hacky.

The other option is to expose the `_cachedFormat` on MutableSettings or
`PerTraceSettings` directly instead, and then have `DatadogLoggingScope`
grab the ambient `CurrentPerTraceSettings.Settings` value. This is a bit
neater in that it removes the chance of ping-pong between logging scope
instances if there are overlapping traces, but it's also a bit meh
because it's moving very integration-specific stuff into somewhere more
"global". But maybe this is just a facility we _should create 🤷‍♂️ What
do people think?

## Test coverage

Added unit tests. They fail before this PR, and pass afterwards

## Other details

Noticed while I was looking for any fallout from #7695