Encode changes to vendored code added in #7884 (#7913)
## Summary of changes

Encodes changes made manually to vendored code in #7884

## Reason for change

In #7884, we manually fixed some problematic code in vendored code. This
PR encodes those changes in the vendors tool, ostensibly so that
re-running the tool recreates the currently vendored code.

Unfortunately, that's not the case. The code that's been vendored for
all of the `System.` libraries _cannot_ be regenerated using the tool.

## Implementation details

Added the transforms, and confirm they work as expected. However, it's a
bit moot as we currently can't regenerate the existing code.

## Test coverage

Manually tested that the new changes were reproduced, then reverted,
because it breaks a bunch of existing vendored code 😅