Mark all classes as `sealed` or `static` (#7879)
## Summary of changes

- Upgrades the `SealedAnalyzer` added in #7878 from `info` to `warning`
for specific projects
- Fix all the errors it flagged

## Reason for change

As described in `#7878`, making types `sealed` can have performance
benefits. It's basically a free small perf bump, and there's not
_really_ any downsides.

## Implementation details

After writing the analyzer, I got claude code to fix all the violations.
Annoyingly, `dotnet format` doesn't seem to work with this analyzer
because it's a "compilation end" analyzer.

Then I went through all the changes and confirmed they're valid. I also
changed a bunch to `static` where they were stateless and effectively
_should_ have been static.

> I initially made all the integrations and aspects `static` instead of
`sealed`, but then realised the integration are used as generic
arguments, and I suspect that's not allowed at runtime

I have written some comments on the PR for places where we had to
explicitly ignore, but GitHub can't handle it, so here's some others 😅

- `Tracer` - had to not marked this sealed, because we derive from it in
tests 🙁 This one is very annoying, I may try to fix in a subsequent PR.
- `DirectSubmissionLogSink` - we were deriving from this in tests so
that we could call a protected method, so just added a way to call that
instead.
- `PerformanceCountersListener` - we derive from this in tests.

## Test coverage

As long as everything passes, we should be good 🤞 

## Other details

Stacked on
- https://github.com/DataDog/dd-trace-dotnet/pull/7878

https://datadoghq.atlassian.net/browse/LANGPLAT-813