Fix usages of `new Tracer()` which are never cleaned up (#7938)
## Summary of changes

Fixes places in tests that currently use `new Tracer()` to use
`ScopedTracer` instead

## Reason for change

When you call `new Tracer()`, a bunch of background jobs are started,
some on thread pool threads, some on dedicated threads. If you don't
explicitly shut down the tracer, these jobs may keep living, and
ultimately act as a drain on resources in CI, leading to flake. e.g. a
check of a recent failing test in CI revealed a lot of threads running
the serialization loop, which potentially points to this kind of issue

<img width="221" height="90" alt="image"
src="https://github.com/user-attachments/assets/f8bc5970-d3d6-4395-968e-cf655b2176c5"
/>

## Implementation details

We introduced `ScopedTracer` and `TracerHelper.Create()` to make doing
this easier, by having a simple async disposable. This PR updates
everywhere that was previously doing `new Tracer()` to use the helpers
instead. There are basically 3 patterns for this:

- If `new Tracer()` was called in a constructor, implement
`IAsyncLifetime` and dispose it in the `DisposeAsync()` call.
- If `new Tracer()` is called inline, use `await using var tracer =
TracerHelper.Create()`
- If a helper is used, find all usages, and ensure they always call
`await using`

## Test coverage

All covered by existing tests, so hopefully we're good 🤞 

## Other details

This seems to be a _relatively_ recent issue in CI, but I'm not sure
why, these usages have persisted for a long time...