Small perf improvement for some `TagList` implementations (#7887)
## Summary of changes

Tags that are readonly should be implemented as such, and return a value
directly

## Reason for change

We have some tags that currently always return the same value, but are
implemented as writeable fields. This requires a backing field, and so
means the object is _sliiiightly_ more allocatey than it needs to be.

## Implementation details

Switch to using ` => ` readonly properties instead of `{ get; set; }`

## Test coverage

If the tests pass, it's fine